資料関連修正

+++ src/main/java/la/bean/AddResultBean.java
@@ -5,7 +5,7 @@ import java.io.Serializable;
 public class AddResultBean implements Serializable{
 
 	private int bookId;
-	private int isbn;
+	private Long isbn;
 	private String title;
 	private int code;
 	private String author;
@@ -15,7 +15,7 @@ public class AddResultBean implements Serializable{
 	
 	public AddResultBean() {}
 	
-	public AddResultBean(int bookId , int isbn , String title , int code , String author ,
+	public AddResultBean(int bookId , Long isbn , String title , int code , String author ,
 			String publicher , String publicationDate , String arrivalDate) {
 		
 		this.bookId = bookId;
@@ -47,10 +47,10 @@ public class AddResultBean implements Serializable{
 		this.bookId = bookId;
 	}
 
-	public int getIsbn() {
+	public Long getIsbn() {
 		return isbn;
 	}
-	public void setIsbn(int isbn) {
+	public void setIsbn(Long isbn) {
 		this.isbn = isbn;
 	}
